Update neopixel_strip.py
something isn't quite right when I try this on my circuit playground express.  it runs without error but does not light up any of the lights. planning on working on this again later.  advice is appreciated but not expected

note: change the board pin variable pin_placement to work with external neopixels
it worked before I changed it to a class and it was using board.D6 for the neopixel strand